Fixed file and checkroutes function
I removed the unnecessary loop in citynamearr and instead accessed the data directly from the City class. From there, I called the data in the Wayfinder class. After that, I created a method named yazdir() to check if the data was being passed correctly. Once I confirmed that the data was being passed correctly, I added functionality in the Wayfinder class to verify whether the cities written in the route section matched the cities listed in the map1.txt file.